+++ dataland-csvconverter/src/main/kotlin/org/dataland/csvconverter/CsvToJsonConverter.kt
@@ -1,0 +1,217 @@
+package org.dataland.csvconverter
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.SerializationFeature
+import com.fasterxml.jackson.dataformat.csv.CsvMapper
+import com.fasterxml.jackson.dataformat.csv.CsvSchema
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import org.dataland.datalandbackend.model.CompanyIdentifier
+import org.dataland.datalandbackend.model.CompanyInformation
+import org.dataland.datalandbackend.model.EuTaxonomyData
+import org.dataland.datalandbackend.model.EuTaxonomyDetailsPerCashFlowType
+import org.dataland.datalandbackend.model.enums.AttestationOptions
+import org.dataland.datalandbackend.model.enums.IdentifierType
+import org.dataland.datalandbackend.model.enums.StockIndex
+import org.dataland.datalandbackend.model.enums.YesNo
+import java.io.File
+import java.io.FileReader
+import java.math.BigDecimal
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+/**
+ * Class to transform company information and EU Taxonomy data delivered by csv into json format
+ * @param filePath location of the csv file to be transformed
+ */
+class CsvToJsonConverter(private val filePath: String) {
+
+    private val objectMapper = ObjectMapper().registerModule(JavaTimeModule())
+        .configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
+    private val notAvailableString = "n/a"
+    private val euroUnitConverter = "1000000"
+
+    private val columnMapping = mapOf(
+        "companyName" to "Company name",
+        "headquarters" to "Headquarter",
+        "sector" to "Sektor",
+        "marketCap" to "Market Capitalization (EURmm)",
+        "reportingDateOfMarketCap" to "Market Capitalization Date",
+        "totalRevenue" to "Total Revenue in EURmio",
+        "totalCapex" to "Total CapEx EURmio",
+        "totalOpex" to "Total OpEx EURmio",
+        "eligibleRevenue" to "Eligible Revenue",
+        "eligibleCapex" to "Eligible CapEx",
+        "eligibleOpex" to "Eligible OpEx",
+        "alignedRevenue" to "Aligned Revenue",
+        "alignedCapex" to "Aligned CapEx",
+        "alignedOpex" to "Aligned OpEx",
+        "companyType" to "IS/FS",
+        "reportObligation" to "NFRD Pflicht",
+        "attestation" to "Assurance"
+    )
+
+    private val stockIndexMapping = mapOf(
+        StockIndex.PrimeStandard to "Prime Standard",
+        StockIndex.GeneralStandard to "General Standard",
+        StockIndex.ScaleHdax to "Scale",
+        StockIndex.Cdax to "CDAX",
+        StockIndex.Gex to "GEX",
+        StockIndex.Dax to "DAX",
+        StockIndex.Mdax to "MDAX",
+        StockIndex.Sdax to "SDAX",
+        StockIndex.TecDax to "TecDAX",
+        StockIndex.Dax50Esg to "DAX 50 ESG"
+    )
+
+    private val identifierMapping = mapOf(
+        IdentifierType.PermId to "PermID",
+        IdentifierType.Isin to "ISIN"
+    )
+
+    private inline fun <reified T> readCsvFile(fileName: String): List<T> {
+        FileReader(fileName).use { reader ->
+            return CsvMapper()
+                .readerFor(T::class.java)
+                .with(CsvSchema.emptySchema().withHeader().withColumnSeparator(';'))
+                .readValues<T>(reader)
+                .readAll()
+                .toList()
+        }
+    }
+
+    private val rawCsvData: List<Map<String, String>> = readCsvFile(filePath)
+
+    private fun getValue(header: String, csvData: Map<String, String>): String {
+        return csvData[header]!!.ifBlank {
+            notAvailableString
+        }
+    }
+
+    private fun getStockIndices(csvLineData: Map<String, String>): List<StockIndex> {
+        return stockIndexMapping.keys.filter { csvLineData[stockIndexMapping[it]]!!.isNotBlank() }
+    }
+
+    private fun getIdentifiers(csvLineData: Map<String, String>): List<CompanyIdentifier> {
+        val identifiers = identifierMapping.keys.map {
+            CompanyIdentifier(identifierValue = getValue(identifierMapping[it]!!, csvLineData), identifierType = it)
+        }
+        return identifiers.filter { it.identifierValue != notAvailableString }
+    }
+
+    private fun getMarketCap(columnHeader: String, csvLineData: Map<String, String>): BigDecimal {
+        return getAbsoluteValue(columnHeader, csvLineData)!!
+    }
+
+    private fun validateLine(csvLineData: Map<String, String>): Boolean {
+        // Skip all lines with financial companies or without market cap
+        return !(
+            getValue(columnMapping["companyType"]!!, csvLineData) in listOf("FS", notAvailableString) ||
+                getValue(columnMapping["marketCap"]!!, csvLineData) == notAvailableString
+            )
+    }
+
+    private fun buildListOfCompanyInformation(): List<CompanyInformation> {
+        return rawCsvData.filter { validateLine(it) }.map {
+            CompanyInformation(
+                companyName = getValue(columnMapping["companyName"]!!, it),
+                headquarters = getValue(columnMapping["headquarters"]!!, it),
+                sector = getValue(columnMapping["sector"]!!, it),
+                marketCap = getMarketCap(columnMapping["marketCap"]!!, it),
+                reportingDateOfMarketCap = LocalDate.parse(
+                    getValue(columnMapping["reportingDateOfMarketCap"]!!, it),
+                    DateTimeFormatter.ofPattern("MM.dd.yyyy")
+                ),
+                identifiers = getIdentifiers(it),
+                indices = getStockIndices(it)
+            )
+        }
+    }
+
+    /**
+     * Method to write the transformed data into json
+     */
+    fun writeJson() {
+        objectMapper.writerWithDefaultPrettyPrinter()
+            .writeValue(File("./CompanyInformation.json"), buildListOfCompanyInformation())
+        objectMapper.writerWithDefaultPrettyPrinter()
+            .writeValue(File("./CompanyAssociatedEuTaxonomyData.json"), buildListOfEuTaxonomyData())
+    }
+
+    private fun getReportingObligation(csvLineData: Map<String, String>): YesNo {
+        return if (getValue(columnMapping["reportObligation"]!!, csvLineData) == "Ja") {
+            YesNo.Yes
+        } else {
+            YesNo.No
+        }
+    }
+
+    private fun getAttestation(csvLineData: Map<String, String>): AttestationOptions {
+        return when (getValue(columnMapping["attestation"]!!, csvLineData)) {
+            "reasonable" -> AttestationOptions.ReasonableAssurance
+            "limited" -> AttestationOptions.LimitedAssurance
+            else -> AttestationOptions.None
+        }
+    }
+
+    private fun getNumericValue(columnHeader: String, csvLineData: Map<String, String>): BigDecimal? {
+        return if (getValue(columnHeader, csvLineData).contains("%")) {
+            getPercentageValue(columnHeader, csvLineData)
+        } else {
+            getAbsoluteValue(columnHeader, csvLineData)
+        }
+    }
+
+    private fun getAbsoluteValue(columnHeader: String, csvData: Map<String, String>): BigDecimal? {
+        // The numeric value conversion assumes the figures to be provided in millions using german notation
+        // , hence, "," as decimal separator and "." to separate thousands
+        return getValue(columnHeader, csvData).trim().replace(".", "").replace(",", ".")
+            .toBigDecimalOrNull()?.multiply(euroUnitConverter.toBigDecimal())
+    }
+
+    private fun getPercentageValue(columnHeader: String, csvData: Map<String, String>): BigDecimal? {
+        return getValue(columnHeader, csvData).trim().replace(".", "").replace(",", ".")
+            .replace("%", "").toBigDecimalOrNull()?.multiply("0.01".toBigDecimal())
+    }
+
+    private fun buildListOfEuTaxonomyData(): List<EuTaxonomyData> {
+        val outputListOfEuTaxonomyData: MutableList<EuTaxonomyData> = mutableListOf()
+        for (csvData in rawCsvData) {
+            if (!validateLine(csvData)) {
+                continue
+            }
+            outputListOfEuTaxonomyData.add(
+                EuTaxonomyData(
+                    reportObligation = getReportingObligation(csvData),
+                    attestation = getAttestation(csvData),
+                    capex = EuTaxonomyDetailsPerCashFlowType(
+                        total = getNumericValue(columnMapping["totalCapex"]!!, csvData),
+                        aligned = getNumericValue(columnMapping["alignedCapex"]!!, csvData),
+                        eligible = getNumericValue(columnMapping["eligibleCapex"]!!, csvData)
+                    ),
+                    opex = EuTaxonomyDetailsPerCashFlowType(
+                        total = getNumericValue(columnMapping["totalOpex"]!!, csvData),
+                        aligned = getNumericValue(columnMapping["alignedOpex"]!!, csvData),
+                        eligible = getNumericValue(columnMapping["eligibleOpex"]!!, csvData)
+                    ),
+                    revenue = EuTaxonomyDetailsPerCashFlowType(
+                        total = getNumericValue(columnMapping["totalRevenue"]!!, csvData),
+                        aligned = getNumericValue(columnMapping["alignedRevenue"]!!, csvData),
+                        eligible = getNumericValue(columnMapping["eligibleRevenue"]!!, csvData)
+                    ),
+                )
+            )
+        }
+        return outputListOfEuTaxonomyData
+    }
+    companion object {
+        /**
+         * The corresponding main class to run the CSV converter. Execute by running:
+         * "./gradlew :dataland-csvconverter:run --args="<FileLocation>"
+         * where <FileLocation> is the location of the CSV file to be converted
+         */
+        @JvmStatic
+        fun main(args: Array<String>) {
+            CsvToJsonConverter(File(args.first()).path).writeJson()
+        }
+    }
+}
