+++ settings.gradle.kts
@@ -1,2 +1,2 @@
 rootProject.name = "dataland"
-include("dataland-backend", "dataland-e2etests", "dataland-frontend", "dataland-prepopulator")
+include("dataland-backend", "dataland-e2etests", "dataland-frontend", "dataland-csvconverter")
